Refactor Handlers into Handlers::Auth and Handlers::Utils, removed unused imports
