+++ src/models/mod.rs
@@ -1,1 +1,1 @@
-pub mod user;
+pub mod models;
