+++ src/services/auth/invite.rs
@@ -1,21 +1,13 @@
 use axum::{
-    http::StatusCode,
-    response::{IntoResponse, Response},
-    Json,
+    http::StatusCode
 };
-use axum_extra::extract::cookie::{Cookie, CookieJar};
-use bcrypt::{hash, verify, DEFAULT_COST};
 use chrono::{DateTime, Duration, Utc};
-use jsonwebtoken::{decode, encode, DecodingKey, EncodingKey, Header, Validation};
-use sqlx::PgPool;
 use uuid::Uuid;
 
-use crate::models::user::{
-    AuthResponse, Claims, CreateAdminRequest, CreateInvitationRequest,
-    Invitation, InvitationResponse, LoginRequest, RegisterWithInvitationRequest,
-    User, UserResponse, SearchUsersQuery, SearchUsersResponse,
+use crate::models::models::{
+    CreateInvitationRequest,
+    Invitation, InvitationResponse
 };
-use crate::services::auth;
 use crate::services::luma::LumaService;
 
 pub async fn create_invitation(
