+++ src/services/auth/signup.rs
@@ -1,21 +1,16 @@
 use axum::{
     http::StatusCode,
-    response::{IntoResponse, Response},
-    Json,
+    response::{ Response}
 };
-use axum_extra::extract::cookie::{Cookie, CookieJar};
-use bcrypt::{hash, verify, DEFAULT_COST};
-use chrono::{DateTime, Duration, Utc};
-use jsonwebtoken::{decode, encode, DecodingKey, EncodingKey, Header, Validation};
-use sqlx::PgPool;
-use uuid::Uuid;
+use axum_extra::extract::cookie::{ CookieJar};
+use bcrypt::{hash, DEFAULT_COST};
+use chrono::{DateTime, Utc};
 
-use crate::models::user::{
-    AuthResponse, Claims, CreateAdminRequest, CreateInvitationRequest,
-    Invitation, InvitationResponse, LoginRequest, RegisterWithInvitationRequest,
-    User, UserResponse, SearchUsersQuery, SearchUsersResponse,
+use crate::models::models::{
+    CreateAdminRequest,
+    Invitation, RegisterWithInvitationRequest, User
 };
-use crate::services::auth;
+
 use crate::services::luma::LumaService;
 
 pub async fn create_admin(
