+++ src/services/auth/utils.rs
@@ -4,18 +4,14 @@ use axum::{
     Json,
 };
 use axum_extra::extract::cookie::{Cookie, CookieJar};
-use bcrypt::{hash, verify, DEFAULT_COST};
 use chrono::{DateTime, Duration, Utc};
 use jsonwebtoken::{decode, encode, DecodingKey, EncodingKey, Header, Validation};
-use sqlx::PgPool;
 use uuid::Uuid;
 
-use crate::models::user::{
-    AuthResponse, Claims, CreateAdminRequest, CreateInvitationRequest,
-    Invitation, InvitationResponse, LoginRequest, RegisterWithInvitationRequest,
-    User, UserResponse, SearchUsersQuery, SearchUsersResponse,
+use crate::models::models::{
+    AuthResponse, Claims,
+    User, UserResponse
 };
-use crate::services::auth;
 use crate::services::luma::LumaService;
 
 pub async fn check_admin_setup(service: &LumaService) -> Result<bool, (StatusCode, String)> {
